requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.4
black==25.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
curl_cffi==0.12.0
Django==5.2.4
djangorestframework==3.16.0
frozendict==2.4.6
idna==3.10
multitasking==0.0.11
mypy_extensions==1.1.0
numpy==2.2.6
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
protobuf==6.31.1
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uv==0.7.10
websockets==15.0.1
yfinance==0.2.65
